+++ vitest.config.ts
@@ -1,4 +1,5 @@
 /// <reference types="vitest" />
+
 import { defineConfig } from 'vitest/config';
 
 export default defineConfig({